requirements.win.txt
aiofiles==23.1.0
aiomysql==0.1.1
aiosqlite==0.17.0
async-timeout==4.0.2
hiredis==2.2.1
html5tagger==1.3.0
httptools==0.5.0
iso8601==1.1.0
multidict==6.0.4
packaging==23.1
PyMySQL==1.0.3
pypika-tortoise==0.1.6
pypinyin==0.48.0
pytz==2023.3
redis==4.5.4
sanic==23.3.0
Sanic-Cors==2.2.0
sanic-redis==0.4.0
sanic-routing==22.8.0
tortoise-orm==0.19.3
tracerite==1.1.0
typing_extensions==4.5.0
ujson==5.7.0
websockets==11.0.2
